refactor: reduce shared module
