+++ src/favorites/favorites.controller.ts
@@ -8,11 +8,9 @@ import {
   Post,
   UseInterceptors,
 } from '@nestjs/common';
-import { ApiPath, ResCode } from 'src/shared/constants/constants';
+import { ApiPath, Entity, ResCode } from 'src/shared/constants/constants';
 import { FavsEntity } from 'src/shared/types/entity-id';
-import { Entity } from 'src/temp-db';
 import { FavoritesService } from './favorites.service';
-import { FavoritesResponse } from './favorites.dto';
 import { Favorites } from './favorites.entity';
 
 @Controller(ApiPath.favorites)
