+++ src/favorites/favorites.module.ts
@@ -1,12 +1,11 @@
-import { forwardRef, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { FavoritesService } from './favorites.service';
 import { FavoritesController } from './favorites.controller';
-import { SharedModule } from 'src/shared/shared.module';
 import { favoritesProviders } from './favorites.providers';
 import { DatabaseModule } from 'src/database/database.module';
 
 @Module({
-  imports: [SharedModule, DatabaseModule],
+  imports: [DatabaseModule],
   providers: [...favoritesProviders, FavoritesService],
   controllers: [FavoritesController],
 })
