--- src/temp-db.ts
@@ -1,7 +1,0 @@
-export enum Entity {
-  artists = 'artists',
-  albums = 'albums',
-  tracks = 'tracks',
-  favorites = 'favorites',
-  users = 'users',
-}
